add bounds steps feature
